Update install script to handle linux or mac os.

+++ tools/install/install_bazel.sh
@@ -26,15 +26,31 @@ fi
 
 set -e
 
-# Install bazel.
-mkdir -p /bazel
-cd /bazel
-if [[ ! -f "bazel-$BAZEL_VERSION-installer-linux-x86_64.sh" ]]; then
-  curl -fSsL -O https://github.com/bazelbuild/bazel/releases/download/$BAZEL_VERSION/bazel-$BAZEL_VERSION-installer-linux-x86_64.sh
-fi
-chmod +x /bazel/bazel-*.sh
-/bazel/bazel-$BAZEL_VERSION-installer-linux-x86_64.sh
-rm -f /bazel/bazel-$BAZEL_VERSION-installer-linux-x86_64.sh
+if [[ "$OSTYPE" == "linux-gnu" ]]; then
+    # Install bazel.
+    mkdir -p /bazel
+    cd /bazel
+    if [[ ! -f "bazel-$BAZEL_VERSION-installer-linux-x86_64.sh" ]]; then
+        curl -fSsL -O https://github.com/bazelbuild/bazel/releases/download/$BAZEL_VERSION/bazel-$BAZEL_VERSION-installer-linux-x86_64.sh
+    fi
+    chmod +x /bazel/bazel-*.sh
+    /bazel/bazel-$BAZEL_VERSION-installer-linux-x86_64.sh
+    rm -f /bazel/bazel-$BAZEL_VERSION-installer-linux-x86_64.sh
 
-# Enable bazel auto completion.
-echo "source /usr/local/lib/bazel/bin/bazel-complete.bash" >> ~/.bashrc
+    # Enable bazel auto completion.
+    grep bazel-complete.bash ~/.bashrc && echo "source /usr/local/lib/bazel/bin/bazel-complete.bash" >> ~/.bashrc
+elif [[ "$OSTYPE" == "darwin"* ]]; then
+        # Mac OSX
+    if [[ $(which brew) == "" ]]; then
+        /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+    fi
+
+    brew install bazel
+    #don't fail if bazel upgrade returns error which it does if we're already at the latest version
+    set +e
+    brew upgrade bazel
+    set -e
+    brew switch bazel $BAZEL_VERSION
+    #brew automatically install bash completion.
+fi
+
